+++ ktlint-ruleset-experimental/src/test/kotlin/com/pinterest/ktlint/ruleset/experimental/FunctionSignatureRuleTest.kt
@@ -1046,6 +1046,36 @@ class FunctionSignatureRuleTest {
             .hasNoLintViolations()
     }
 
+    @Test
+    fun `Issue 1690 - Given a function preceded by an annotation array`() {
+        val code =
+            """
+            // $MAX_LINE_LENGTH_MARKER                   $EOL_CHAR
+            internal fun foo1(foo1: Foo, foo2: Foo): Foo =
+                "foooooooooooooooooooooooooooooooooooooo"
+
+            @Bar
+            internal fun foo2(foo1: Foo, foo2: Foo): Foo =
+                "foooooooooooooooooooooooooooooooooooooo"
+
+            @[Bar]
+            internal fun foo2(foo1: Foo, foo2: Foo): Foo =
+                "foooooooooooooooooooooooooooooooooooooo"
+
+            @[Bar1 Bar2 Bar3 Bar4 Bar5 Bar6 Bar7 Bar8 Bar9]
+            internal fun foo2(foo1: Foo, foo2: Foo): Foo =
+                "foooooooooooooooooooooooooooooooooooooo"
+
+            @[Bar1 // some comment
+            Bar2]
+            internal fun foo2(foo1: Foo, foo2: Foo): Foo =
+                "foooooooooooooooooooooooooooooooooooooo"
+            """.trimIndent()
+        functionSignatureWrappingRuleAssertThat(code)
+            .setMaxLineLength()
+            .hasNoLintViolations()
+    }
+
     private companion object {
         const val EOL_CHAR = '#'
         const val UNEXPECTED_SPACES = "  "